# Fix ssao only sampling mip 0