@JMS55 2024-01-12
Makes no difference on a 3080 lol, but don't know how to run a profiler on my laptop /shrug